+++ test/FormElementManagerFactoryTest.php
@@ -10,6 +10,7 @@ namespace ZendTest\Form;
 use Interop\Container\ContainerInterface;
 use PHPUnit\Framework\TestCase;
 use Zend\Form\ElementInterface;
+use Zend\Form\Element\Number;
 use Zend\Form\FormElementManager;
 use Zend\Form\FormElementManagerFactory;
 use Zend\ServiceManager\ServiceLocatorInterface;
@@ -69,5 +70,101 @@ class FormElementManagerFactoryTest extends TestCase
 
         $elements = $factory->createService($container->reveal());
         $this->assertSame($element, $elements->get('test'));
+    }
+
+    public function testConfiguresFormElementsServicesWhenFound()
+    {
+        $element = $this->prophesize(ElementInterface::class)->reveal();
+        $config = [
+            'form_elements' => [
+                'aliases' => [
+                    'test' => Number::class,
+                ],
+                'factories' => [
+                    'test-too' => function ($container) use ($element) {
+                        return $element;
+                    },
+                ],
+            ],
+        ];
+
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(false);
+        $container->has('config')->willReturn(true);
+        $container->get('config')->willReturn($config);
+
+        $factory = new FormElementManagerFactory();
+        $elements = $factory($container->reveal(), 'FormElementManager');
+
+        $this->assertInstanceOf(FormElementManager::class, $elements);
+        $this->assertTrue($elements->has('test'));
+        $this->assertInstanceOf(Number::class, $elements->get('test'));
+        $this->assertTrue($elements->has('test-too'));
+        $this->assertSame($element, $elements->get('test-too'));
+    }
+
+    public function testDoesNotConfigureFormElementsServicesWhenServiceListenerPresent()
+    {
+        $element = $this->prophesize(ElementInterface::class)->reveal();
+        $config = [
+            'form_elements' => [
+                'aliases' => [
+                    'test' => Number::class,
+                ],
+                'factories' => [
+                    'test-too' => function ($container) use ($element) {
+                        return $element;
+                    },
+                ],
+            ],
+        ];
+
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(true);
+        $container->has('config')->shouldNotBeCalled();
+        $container->get('config')->shouldNotBeCalled();
+
+        $factory = new FormElementManagerFactory();
+        $elements = $factory($container->reveal(), 'FormElementManager');
+
+        $this->assertInstanceOf(FormElementManager::class, $elements);
+        $this->assertFalse($elements->has('test'));
+        $this->assertFalse($elements->has('test-too'));
+    }
+
+    public function testDoesNotConfigureFormElementsServicesWhenConfigServiceNotPresent()
+    {
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(false);
+        $container->has('config')->willReturn(false);
+        $container->get('config')->shouldNotBeCalled();
+
+        $factory = new FormElementManagerFactory();
+        $elements = $factory($container->reveal(), 'FormElementManager');
+
+        $this->assertInstanceOf(FormElementManager::class, $elements);
+    }
+
+    public function testDoesNotConfigureFormElementServicesWhenConfigServiceDoesNotContainFormElementsConfig()
+    {
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(false);
+        $container->has('config')->willReturn(true);
+        $container->get('config')->willReturn(['foo' => 'bar']);
+        $container->has('MvcTranslator')->willReturn(false); // necessary due to default initializers
+
+        $factory = new FormElementManagerFactory();
+        $elements = $factory($container->reveal(), 'FormElementManager');
+
+        $this->assertInstanceOf(FormElementManager::class, $elements);
+        $this->assertFalse($elements->has('foo'));
     }
 }